+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Config option added: ckanext.downloadall.dataset_fields_to_add_to_datapackage for including custom fields from the dataset in the datapackage.json
+
 ## [0.0.2] - 2019-06-30
 ### Added
 - Command-line interface.